SafeFlights
There are N islands (numbered from 0 to N-1). You have a list of all flights that are going to take place. Each flight will be given as pair of numbers (A and B). This means that there will be a plane which constantly flies from A to B and back. The flight is safe if there isn't another flight between A and B. Your task is to calculate the number of distinct island-to-island routes which are safe to travel.